requirements.txt
py-clob-client>=0.17.0
web3>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
